Permit `"same_kind"` casting for element-wise in-place operators (#2170)
The PR proposes to permit `"same_kind"` casting for element-wise
in-place operators. The implementation leverages on dpctl changes added
in scope of [PR#1827](https://github.com/IntelPython/dpctl/pull/1827).

It also adds callbacks to support in-place bit-wise operators (leverages
on dpctl changes from
[RR#1447](https://github.com/IntelPython/dpctl/pull/1447)).

The PR removes a temporary workaround from `dpnp.wrap` which depends on
the implemented changes.

+++ dpnp/dpnp_array.py
@@ -399,7 +399,7 @@ class dpnp_array:
             axes = [(-2, -1), (-2, -1), (-2, -1)]
 
         try:
-            dpnp.matmul(self, other, out=self, axes=axes)
+            dpnp.matmul(self, other, out=self, dtype=self.dtype, axes=axes)
         except AxisError:
             # AxisError should indicate that the axes argument didn't work out
             # which should mean the second operand not being 2 dimensional.
